src/FE/requirements.txt
streamlit==1.28.1
requests==2.32.0
python-dotenv==1.0.0